Changed the way VR Controls worked.  Before I had each of the different attacks as a hard-coded member of the VR Controls class that was dragged in in the inspector.  Now there is just a master list of attacks that gets dragged into VR Controls as an array.  There is now an AttackTypes enum that every AAttack has as a public field.  Attacks are changed by iterating through the master array and looking for an AAttack that matches the desired enum.  An input type enum has also been added for each of the controls.
This allows me to have each of the attacks in the attack choice UI to just select one of the attack type enums and send that up to AttackSelection that also keeps track of the current input type.  The new attack and its corresponding input is then sent to VR Controls where it is changed.